==> src/scRNAseq/scripts/R/scRNA-seq/2.batchRemove.harmony.R <==
library(Seurat)
library(harmony)
library(patchwork)
options(bitmapType = "cairo")
# don't need split Seurat
SC = readRDS("/home/lsg/Data/glioblastoma/output/rds/QC/SC_QC.rds")
# TE = readRDS("/home/lsg/Data/glioblastoma/output/rds/merge/TE.rds")
# GBM = readRDS("/home/lsg/Data/glioblastoma/output/rds/QC/GBM_QC.rds")
harmony = function(Sat,out){
    # 检查数据中的极端值或缺失值
    print(summary(Sat@assays$RNA@counts))
    cat(sum(is.nan(Sat@assays$RNA@counts)),sum(!is.finite(Sat@assays$RNA@counts)) )
    Sat = NormalizeData(Sat,normalization.method = "LogNormalize") 
    Sat = FindVariableFeatures(Sat,selection.method = "vst",nfeatures = 1000) 
    print(length(VariableFeatures(object = Sat)))
    Sat = ScaleData(Sat,features = rownames(Sat)) 
    Sat = RunPCA(Sat,features = VariableFeatures(object = Sat),reduction.name = "pca")
    p = ElbowPlot(Sat,ndims = 50)
    jpeg(width = 150, height = 150, units = "mm", res = 500, file = paste("./figures/R/",out,"_Dim_plot.png",sep=""))
        print(p)
    dev.off()
    # write.csv(Sat@reductions$pca@cell.embeddings,file = "/home/lsg/Data/glioblastoma/figures/R/SC_harmony_pca_embeddings.csv")
    # print(ncol(Sat@reductions$pca@cell.embeddings))
    # # don't remove batch
    # Sat = RunUMAP(Sat,reduction = "pca",dims = 1:10,reduction.name = "umap_naive")
    # print("----------1--------------")
    # #using harmony to remove batch;dims=1:30->select how much pcas to RunUMAP
    # Sat = RunHarmony(Sat, reduction = "pca",group.by.vars = "sample",reduction.save = "harmony")
    # print("----------2--------------")
    # print(ncol(Sat@reductions$harmony@cell.embeddings))
    # Sat = RunUMAP(Sat,reduction = "harmony",dims = 1:30,reduction.name = "umap") 
    # print("----------3--------------")
    # p1 = DimPlot(Sat, reduction = "umap_naive",group.by = "sample")
    # p2 = DimPlot(Sat,reduction = "umap",group.by = "sample")
    # jpeg(width = 150, height = 150, units = "mm", res = 500, file = paste("./figures/R/",out,"_Dim_plot.png",sep=""))
    #     print(p1+p2)
    # dev.off()
    # saveRDS(Sat, file = paste("/home/lsg/Data/glioblastoma/output/rds/",out,"_harmony.rds"))
    return (Sat)
}
# SC = harmony(SC,"SC")
# GBM = harmony(GBM,"GBM")
# print(str(SC))
TE = harmony(SC,"SC")
# GBM = harmony(GBM)
